fixed scripts, mainly paths

+++ scripts/get_coco_dataset_2017.sh
@@ -2,43 +2,27 @@
 
 ### Recommended to run 'nohup ./<this_script> &' to prevent interruption from SSH session termination.
 
-wait_to_finish() {
-    for pid in "${download_pids[@]}"; do
-        while kill -0 "$pid"; do
-            sleep 30
-        done
-    done
-}
-
-
 # Update for default OS specific package manager.
 # sudo yum -y install java-1.8.0
 # sudo yum -y remove java-1.7.0-openjdk
 
-mkdir -p coco/images/ coco/annotations/
-
-download_pids=()
+mkdir -p ../data/dataset/coco/images/ ../data/dataset/coco/annotations/
 
 ### 2017 COCO Dataset ###
 
 echo "Downloading COCO dataset..."
-curl -OL "http://images.cocodataset.org/zips/val2017.zip" &
-download_pids+=("$!")
-curl -OL "http://images.cocodataset.org/annotations/annotations_trainval2017.zip" &
-download_pids+=("$!")
-
-wait_to_finish download_pids
-
-inflate_pids=()
+curl -OL "http://images.cocodataset.org/zips/val2017.zip"  
+curl -OL "http://images.cocodataset.org/annotations/annotations_trainval2017.zip"
 
 echo "Extracting..."
 
-unzip 'val2017.zip' -d coco/images/ &
-inflate_pids+=("$!")
-unzip 'annotations_trainval2017.zip' -d coco/annotations/ & # Inflates to 'coco/annotations'.
-inflate_pids+=("$!")
+unzip 'val2017.zip' -d ../data/dataset/coco/images/
+unzip 'annotations_trainval2017.zip' -d ../data/dataset/coco/ # Inflates to '/annotations'.
 
-wait_to_finish inflate_pids
+if $? -ne 0:
+    echo "Unzipping wasn't successful. Aborting..."
+    return 1
 
 echo "Cleaning up zip files"
+
 rm 'val2017.zip' 'annotations_trainval2017.zip'
